Commit the re-regenerated bridges.
This commit replaces the OWL versions of the "old" bridges (as generated
during the last release process) by versions generated by the new
pipeline.

The OBO versions of the old bridges are *not* replaced, as the new
pipeline creates OWL files directly and does not automatically create
OBO versions -- they are only generated on demand.