Use the latest version of the SSSOM plugin.
There are some cases in Uberon where a same foreign term is referenced
in two xref annotations: one on a valid Uberon term, and one on a
deprecated Uberon term (the latter should probably have been removed
when the term was deprecated, but his has not always been done).

In this situation, it is clear that the xref on the deprecated term
should be ignored and that only the one on the valid term should be
taken into account. This is what version 0.4.3 of the SSSOM plugin does.