Unsat fix: Remove 'pericardial cavity present_in_taxon some Ciona'.
The pericardial cavity, like all terms related to 'heart', is strictly
defined as vertebrate-specific in Uberon. Ciona do have something that
is akin to a pericardium, but since they are not vertebrates, we should
either:

* broaden our heart-related terms (at least 'pericardium', and all terms
  connected to it) so that it is no longer vertebrate-specific;
* create another term to represent pericardium-like structure in
  non-vertebrates.

Until a decision is made between these two approaches, here we simply
remove the statement that the pericardial cavity exists in Ciona, since
it directly violates the vertebrate-specific constraint.