Force refreshing the CL and ZFA mirrors when building the bridges.
Building the bridges requires the CL and ZFA mirrors to be locally
available (we need to extract cross-references from them). But the
bridges are typically generated (e.g. during a release run) with
IMP=false, which prevents the mirror from being downloaded (and will
therefore result in a build failure unless the mirrors are present
because they have been downloaded as part of a previous workflow).