Remove bogus axiom from the local EHDAA2 mirror.
The local EHDAA2 mirror, used to build composite-metazoan, contains the
following axiom:

'extra embryonic coelum' SubClassOf: 'attached to' some 'coelomic cavity
lumen'

This is wrong as 'attached to' (RO:0002371) is range- and
domain-restricted to anatomical structures, while both 'extra embryonic
coelum' (EHDAA2:0004739) and 'coelomic cavity lumen' (EHDAA2:0000267)
are anatomical spaces (immaterial entities).

This should really be fixed in EHDAA2, but the ontology is officially
marked as inactive and has not seen any really meaningful edit for the
past 8 years.