Unmap 'serous sac' from EMAPA:16060 and MA:0000005.
'serous sac' (UBERON:0005906) is mapped with EMAPA:16060. This is wrong,
because the Uberon term represents an entire organ, made of a serous
membrane surrounding an anatomical cavity, while the EMAPA term refers
to a body cavity OR the lining that surrounds it. The Uberon term is
necessarily a material entity, while the EMAPA term is not committed to
being either a material entity or an immaterial entity.

Mapping the two terms together results, when building
composite-metazoan, in the EMAPA term becoming a subclass of the Uberon
term, thereby forcing the EMAPA term to be a material entity. This
causes several unsatisfiability issues because there are other Uberon
terms that are mapped to subclasses of the EMAPA term but represent
immaterial entities.

Likewise for MA:0000005.