Fix composite unsats (#3099)
* Remove old bogus axiom from the local CTENO mirror.

The local CTENO mirror, used to produce composite-metazoan, contains
several axioms coming from an old version of Uberon (CTENO has not been
updated since 2016). One such axiom is

'surface groove' SubClassOf: 'surface structure'

which *may* have been correct in 2016 but is clearly bogus now, since
'surface groove' (UBERON:0006846) is an immaterial anatomical entity
whereas 'surface structure' (UBERON:0003102) is an anatomical structure
(and thus a material anatomical entity).

That single axiom is causing massive unsatisfiability issues when
building composite-metazoan.

* Remove bogus axiom from the local EHDAA2 mirror.

The local EHDAA2 mirror, used to build composite-metazoan, contains the
following axiom:

'extra embryonic coelum' SubClassOf: 'attached to' some 'coelomic cavity
lumen'

This is wrong as 'attached to' (RO:0002371) is range- and
domain-restricted to anatomical structures, while both 'extra embryonic
coelum' (EHDAA2:0004739) and 'coelomic cavity lumen' (EHDAA2:0000267)
are anatomical spaces (immaterial entities).

This should really be fixed in EHDAA2, but the ontology is officially
marked as inactive and has not seen any really meaningful edit for the
past 8 years.

* Unmap 'serous sac' from EMAPA:16060 and MA:0000005.

'serous sac' (UBERON:0005906) is mapped with EMAPA:16060. This is wrong,
because the Uberon term represents an entire organ, made of a serous
membrane surrounding an anatomical cavity, while the EMAPA term refers
to a body cavity OR the lining that surrounds it. The Uberon term is
necessarily a material entity, while the EMAPA term is not committed to
being either a material entity or an immaterial entity.

Mapping the two terms together results, when building
composite-metazoan, in the EMAPA term becoming a subclass of the Uberon
term, thereby forcing the EMAPA term to be a material entity. This
causes several unsatisfiability issues because there are other Uberon
terms that are mapped to subclasses of the EMAPA term but represent
immaterial entities.

Likewise for MA:0000005.