Unsat fix: Move misplace 'present in taxon' assertion.
'skull' is said to be 'present in taxon' some Myxinidae. But Myxinidae
are jawless vertebrates while skull, as defined in Uberon, refers to the
association of a cranium with a mandible. Without a mandible, Myxinidae
have a cranium but no skull. So we move the 'present in taxon' some
Myxinidae axiom from 'skull' to 'cranium'.