Forcefully fix externally caused TC violations.
Remove the mapping between UBERON:0000435 and DMBA:16271 from the DMBA
bridge. Remove from the hra_subset component the assertion that
UBERON:8440041 is present_in_taxon some NCBITaxon:9606.

The mapping and the assertion are bogus and violate the taxon
constraints. They need to be fixed upstream (both the DMBA bridge and
the HRA subset are maintained elsewhere), but until that happens we
remove them from our local copies.