Fetch external mapping sets in a separate phase.
Do not extract the mappings from CL at the same time as we extract the
mappings from Uberon. They are maintained in an external repository (the
CL repository), so they should be treated differently, downloaded
separately and committed to the repository before being used in
downstream rules (instead of being written to a temporary file).

Same for the mappings extracted from ZFA.