Use OWL as the default format for the custom bridges.
All bridges are now generated in the OWL format, and the custom bridges
are provided by upstream already in the OWL format, so we can use them
directly in that format. We do not generate OBO versions by default. If
we do want an OBO version of them, such a version can be generated using
the default rule to produce an OBO bridge from an OWL bridge, there is
no need to have specific rules for the custom bridges.