Update the SSSOM plugin.
Update to the latest version of the SSSOM plugin. This fixes a bug with
the mapping cardinality filter that caused a large part of the CL-ZFA
mappings to be ignored.

We also use the new features of that version to ensure that the
generated mapping sets have a license and an ID.