Fix the custom rules for downloading mirrors.
The custom rules to download mirror (those that are not managed by the
ODK) erroneously and constantly consider that any freshly downloaded
mirror is different from any previously available mirror, because the
previously available is overwritten in the process.

This originates from a known bug in the ODK which has since been fixed.
We apply the same fix here.