Simplify treatment of "local imports".
The "local-" imports (those used to build the composite-* products) do
not actually need much special treatment, certainly not as much as they
used to need.

Here, the only treatment we constantly apply to all the local imports is
to strip axioms that are about terms outside of the foreign ontology's
namespace (in effect, making the import a "base"). A handful of
ontologies still need some extra care to translate old-style properties
into their modern RO equivalents. We do that using `robot rename`
instead of `owltools --rename-entity`.

We do *not* strip disjointness axioms at this step. This will be done
later, at the beginning on the composite pipeline.