Allow quickly refreshing all "local" imports.
Add a new Make target called "all_local_imports". Similarly to the ODK
standard "all_imports", it allows to refresh in one step all the
non-ODK-managed imports.

That target is called by the 'refresh-external-resources' target, so
that local imports can be refreshed at the same time as the normal
imports and the bridges.