Add title and description to the composite- products.
For each composite- product, add a -hdr file containing basic
annotations (title and description) to be used as the first ontology to
be merged when assembling the product. This has the effect of:

a) ensuring the composite product has a descriptive title and
description that match its actual contents;
b) remove all the Uberon-specific top-level ontology annotations.