Remove rules to create uberon-nif-merged.
The uberon-nif-merged.owl product has been impossible to build for (at
least) several months (possibly much more than that) because the link we
try to download the "NIF Gross Anatomy" from is dead.

Besides, the usefulness of this product, even if we could build it, is
dubious since the latest available version of the "NIF Gross Anatomy"
seems to only contain deprecated terms that have been replaced by Uberon
terms.

Adding to that the fact that nobody raised a voice against the removal
of the uberon-nif-merged product since such a removal was evoked
(#3018), there is no obvious reason to keep in place the rules to create
this product. So we remove them here.

closes #3018