Make sure all cross-references use CURIEs.
When generating the mappings component, make sure all cross-references
are only ever written in CURIE form. We do that by:

* using the new option --error-on-unshortenable-iris, to force the
  command to error out if any IRI cannot be shortened for lack of a
  suitable prefix declaration;
* declare the "http://id.nlm.nih.gov/mesh/" prefix so that any
  identifier in that namespace can be shortened;
* rename subject IDs that use the bogus prefix
  "https://meshb.nlm.nih.gov/record/ui?ui=" to force them to use the
  correct prefix.

In the future, the last step should ideally be performed earlier in the
pipeline, just after fetching the Biomappings set and saving it to the
mappings directory. But this will require the "sssom-cli" tool which is
not part of the ODK yet.