Add LICENSE file.
This commit adds the full text of the Creative Commons Attribution 3.0
Unported license in a new LICENSE file at the root of the repository.

This does NOT change the license used by Uberon. Uberon IS already using
the CC BY 3.0 Unported license right now (as indicated by an annotation
in the -edit file). This commit simply makes the license more obvious.