Allow to keep disjointness axioms when building composite products.
All composite-* products are built by, among other operations, stripping
any disjointness axiom in the merged ontology. This is necessary as most
composite-* products would be impossible to build otherwise (the
merge-species step requires the ontology to be consistent, and most of
the collected-* merges are not because of incompatibilities between
Uberon and one or several of the taxon-specific ontologies).

This will not change anytime soon, but we should still at least *aim* to
achieve inter-ontology consistency in the (far) future. This has already
been achieved at least for the composite-drosophila product (Uberon + CL
+ FBbt), so it _is_ possible.

This commit makes the removal of the disjointness axioms optional, so
one can easily try to build a composite-* product with the disjointness
axioms still in place. Simply run the pipeline with
COMPOSITE_STRIP_DISJOINTS=false, e.g.

  sh run.sh make composite-mouse.owl COMPOSITE_STRIP_DISJOINTS=false

For most products this is likely to fail, but this gives us an easy way
to _test_ building with disjointness axioms in place, so that we can
start addressing the inconsistency issues.