Change extension of disjoint_union_over component.
Rename the `disjoint_union_over` component to give it a `.owl`
extension. Other components written in OWL functional syntax uses the
`.owl` extension, no reason to single that component out. The ontology
IRI inside that component is itself using a `.owl` extension, so let's
align the filename with the IRI.