Fix classification of 'insect dorsal ectoderm derivative'.
The term 'insect dorsal ectoderm derivative' should not be classified as
a 'multicellular anatomical structure', as there can be (and indeed,
there are) derivates of the dorsal ectoderm that are single cells (such
as histoblasts) or even acellular structures (such as apodemes).

We classify it as am 'ectoderm-derived structure' instead.

We also add a definition that was missing for that term.

closes #3190