Obsolete 'telencephalic tracts'.
The term 'telencephalic tracts' was created as a generalised version of
the ZFA term 'telencephalic white matter' (defined as "CNS white matter
that is part of the telencephalon"). It seems redundant with 'white
matter of telencephalon' (defined as "a portion of white matter that is
part of a telencephalon").

We obsolete 'telencephalic tracts' and merge it with 'white matter of
telencephalon'.

closes #3160