Obsolete 'common crus'.
The term 'common crus' was created as a generalised version of the
eponym ZFA term (defined as the "duct that forms when the nonampullary
ends of the posterior and anterior semicircular canals merge"). It seems
redundant with 'common crus of semicircular ducts' (defined as "a
nonampullary end of the superior or posterior semicircular ducts").
There does not seem to be any other "common crus" than those of the
semicircular ducts.

We obsolete 'common crus' and merge it into 'common crus of semicircular
ducts'.

closes #3161