Use ODK-managed ROBOT plugins.
The SSSOM plugin for ROBOT is bundled with the ODK, so we no longer need
to manually download it.

As for the Uberon-specific plugin, we still need to download it but we
can rely on the ODK to do that for us--all we need is to declare the
need to use that plugin.