Move the mapping sets from src/ontology/mappings to src/mappings.
We had started to put the mapping sets in src/ontology/mappings, but the
directory recommended by the ODK is src/mappings. So we move the
existing sets to that directory.

We also change the naming convention to remove the `-mappings` suffix
before the extension (i.e. `fbbt.sssom.tsv` instead of
`fbbt-mappings.sssom.tsv`). Again, this is to align with the ODK
recommendations.