Update Makefile with new location of mapping files.
We update the custom Makefile to take into account that the mapping sets
are now in src/mappings instead of src/ontology/mappings.

We also move the rules that generate the sets a bit, so that they are
all in the same section (previously the rule that generated the set
extracted from Uberon's xref was in the bridge section).

We also introduce the "meta" Uberon set, which is a compilation of

* uberon-local (the set derived from Uberon's xrefs);
* cl (the set derived from CL's xrefs);
* zfa (the set derived from ZFA's xrefs);
* fbbt (the set mirrored from FBbt).