Do not commit the locally built mapping sets.
Avoid committing the `uberon-local.sssom.tsv` mapping set to the
repository. This is not needed for now as the set depends only on
Uberon's -edit file, so it can always be re-generated whenever it is
needed even from a freshly cloned repository.

Also avoid committing the "meta" set `uberon.sssom.tsv`. It is merely a
compilation of the other mapping sets (including remote sets that *are*
committed to the repository) and so can also always be re-generated as
needed. In addition, the set is published as a release artifact, which
is how it should be accessed.