Mention ROBOT in the acknowledgements.
If we take the time to mention the OWL API and OWLTools in the
acknowledgements, it is only fair to also mention ROBOT, which is more
and more taking the place of OWLTools.

Consequently, I believe it is also only fair to mention James Overton,
as the main ROBOT developer.