Mention the SSSOM mapping set in the downloadable products.
Make sure the "meta" mapping set, grouping all the mappings used by
Uberon, is listed on the release page. The file is not *currently*
available because we have yet to make a release since we have added the
code to publish it, but it will be upon the next release.