Use CL-provided mapping set.
This commit updates the mappings/bridges pipeline to use the mapping set
provided directly by CL, so that we no longer have to deal with the CL
mappings in Uberon.

We no longer need a ZFA mapping set, since that set only contained the
CL-to-ZFA mappings and those mappings are now provided by CL itself. The
Uberon-to-ZFA mappings are in the uberon-local set, since Uberon xrefs
are the source of truth for those mappings.

When we mirror the FBbt set, we no longer need to keep the FBbt-to-CL
mappings, since again they are in the CL-provided set.

Lastly, the Uberon "meta" set (the one that ultimately gets published)
will now only contain the Uberon-related mappings (Uberon-to-anything)
and will no longer bundle the CL mappings. Basically, the idea is that
from now on every ontology is only responsible for providing its own
mappings:

* FBbt provides the FBbt-to-anything mappings;
* CL provides the CL-to-anything mappings;
* Uberon provides the Uberon-to-anything mappings.

This does not change anything for the *bridges*. Uberon will still
provide both the Uberon-to-anything bridges *and* the CL-to-anything
bridges, because I don't think duplicating the logic that produces those
bridges across both the Uberon and CL repositories would be worth it,
especially since it is in Uberon that those bridges are actually needed
(to build the collected/composite ontologies).