Test release Uberon (not to be merged!)
If you see this commit in the git history someone made a mistake, it was supposed to be for review purpose only.

This PR is the consequence of running the `make uberon` pipeline, omitting the faulty `DMBA:16271` mapping which causes an inconsisteny. All release files are updated, but there are some interesting changes:

- Lots of prefix changes
- The consider axioms are changed to PURL syntax in OWL serialisation? that seems odd

OLD:

<oboInOwl:consider>EFO:0001015</oboInOwl:consider>

New:

<oboInOwl:consider rdf:resource="http://purl.obolibrary.org/obo/EFO_0001015"/>