Make sure OMO:0002000 is NOT stripped from imports.
Since version 1.5, the ODK import system strips by default almost all
annotation properties from the imported modules, except rdfs:label and
IAO:0000115. This is intended to reduce the size of the import modules,
on the rationale that most annotations are "not important" anyway.

Unfortunately, this results in the OMO:0002000 annotation property being
stripped from the RO import, and that property is definitely not a "not
important" one, since it defines the axioms that ROBOT's `expand`
command should inject into the ontology wherever that annotation is
used.

In particular, the expansion of RO:0002175 (`present_in_taxon`) is a
critical step of the check for taxon constraint violations. As a result
of OMO:0002000 being stripped, that expansion is not performed, and the
taxon constraints check is inoperant -- it has been inoperant ever since
we migrated to ODK 1.5, back in March 2024!

This commit explicitly declares which annotation properties should be
preserved by the ODK when refreshing the import modules, instead of
relying on the ODK default (rdfs:label and IAO:0000115), to make sure
that OMO:0002000 is among them.