Remove the tee command.
The added '| tee' at the end of the command line may be what caused the
step to be considered a success even though the command failed.

That is because under the shell -e regime, a command that exits with an
error status does NOT cause the shell to error out if the failing
command is in the middle of a pipeline. That is:

  cmd1 && cmd2 && cmd3

will NOT cause the shell to error out if cmd1 or cmd2 returns an error;
only if cmd3 returns an error will the shell itself error out.

The same logic applies for

  cmd1 && cmd2 | tee file

cmd2 is NOT the last command of the pipeline, so it will not cause the
shell to error out upon failure.