Use ODK 1.5.2 for the CI test suite.
ROBOT 1.9.6 (and its 0.6 ELK reasoner) can catch TC violations that
ROBOT 1.9.5 does not detect.

If Uberon releases are manually performed using the latest ODK (1.5.2,
which includes ROBOT 1.9.6), then it means that we could run at release
time into TC violations that were undetected during the normal
development cycle.

To avoid that, this commit updates the QC workflow to make it use the
latest point release of the ODK.