Fix unsatisfiability of UBERON:0001250.
'red pulp of spleen' violates a taxon constraint because it is
present_in_taxon some 'Danio rerio' but is linked (through 'spleen
primordium', from which the spleen develops) to stomach, which is
never_in_taxon some 'Danio rerio'.

We fix that by removing the following axiom:

  'spleen primordium' part_of some 'dorsal mesogastrium'

That axiom is what links the existence of the spleen to the existence of
the stomach, and is probably too strong a claim. If a spleen can exist
in species without a stomach (as is the case in zebrafish), then the
spleen primordium cannot always be part of a component of the stomach.