Fix incompatibility with ZFA's "head".
In ZFA, the "head" (ZFA:0001114) corresponds to what Uberon calls the
"craniocervical region" (UBERON:0007811), comprising the head proper and
the neck. So we remap ZFA:0001114 to UBERON:0007811, rather than
UBERON:0000033).

Also, the axiom stating that 'pharyngeal arch' (UBERON:0002539) is part
of the 'embryonic head' (UBERON:0008816) is bogus, as some pharyngeal
arches are located in the neck rather than the head. So we create a new
term 'embryonic craniocervical region' to be the embryonic counterpart
of 'craniocervical region' (as 'embryonic head' is the embryonic
counterpart of 'head'), and rewrite the aforementioned axiom to state
that 'pharyngeal arch' is part of the 'embryonic craniocervical region'.

Combined together, these two changes fix about 80 unsatisfiable classes
in the merge product of Uberon and ZFA.

closes #379