Remap ZFA's 'griseum centrale'.
ZFA's 'griseum centrale' (ZFA:0000518) is mapped to Uberon's 'central
gray substance of midbrain' (UBERON:0003040).

This is likely incorrect as the ZFA term refers to a structure that is
part of the hindbrain, whereas the Uberon term refers to a midbrain
structure.

The correct mapping is most likely 'central gray substance of medulla'
(UBERON:0002867), which is part of the hindbrain.

partially addresses #3353