Remap ZFA's 'coelom'.
ZFA's 'coelom' (ZFA:0001438) is mapped to Uberon's 'coelemic cavity
lumen' (UBERON:0002323).

This is incorrect as the ZFA terms refers to a multi-tissue structure
and therefore to a material entity, whereas the Uberon term refers to
the anatomical cavity *within* the coelom (an immaterial entity).

Uberon has another term to refer to the (material) anatomical structure
that forms the coelemic cavity: 'coelom' (UBERON:0011997), the ZFA term
should be mapped to that term instead.

partially addresses #3353