Refactor the rules to create taxon-specific subsets.
The custom Makefile contains two sets of rules to create taxon subsets
in two different ways:

* one set using OWLTools' `--make-species-subset` command (resulting in
  the `*-basic.owl` subsets);
* one set using the files in `src/ontology/contexts` to do basically the
  same thing as `--make-species-subset`, merely in a slightly different
  way (resulting in the `*-view.owl` subsets).

This command replaces both sets of rules by a single rule that relies on
the newly available `create-species-subset` command in the Uberon ROBOT
plugin.

In addition, a new rule is added to allow the creation of a component
file that contains `oboInOwl:inSubset` annotations to tag all the
classes that belong to a given subset. That rule is currently unused,
but the expectation is that it could be used by downstream applications
to facilitate the use of taxon-specific subsets.