Merge some taxon subset tags in the main release products.
In addition to producing taxon subset files, we also want to include,
directly into the release products, the oboInOwl:inSubset annotations
that mark all terms that belong to the taxon subsets.

There is a bit of a conundrum here as the taxon subsets are computed on
the main uberon.owl product, but we want to include them in that very
product. To solve this, we introduce a "near-final" intermediate product
(tmp/uberon.owl, referenced by the new Make variable POSTPROCESS_SRC).
The pipeline that produced the final uberon.owl now produces that
intermediate tmp/uberon.owl, from which the taxon subsets are derived.
The new final uberon.owl is produced simply by merging the intermediate
tmp/uberon.owl with the taxon subset tag files.

Some of the other parts of the Uberon pipeline that were using the final
uberon.owl are now using the intermediate tmp/uberon.owl, because the
taxon subset annotations are not needed for those steps. This notably
concerns the bridge checks and the building of Composite Metazoan.