Do not inject uberon-full.owl into uberon.owl.
The infortunate use of $^ in the rule that produces `uberon.owl` leads
to `uberon-full.owl` being forcefully injected into `uberon.owl`,
because `uberon-full.owl` is declared (in the ODK-generated Makefile) as
a dependency of `uberon.owl`. We must avoid using that variable and only
use the dependencies that are explicitly listed in uberon.Makefile.