Remove redundant injections about RO terms.
We are injecting several statements about RO terms that are redundant
because they correspond to axioms that are already present in upstream
RO.

Here, we remove all such injections. We also prevent the renaming of RO
terms (e.g., avoid renaming 'present in taxon' into 'present_in_taxon').

Of note, the injections on RO:0002492 ('existence ends during') were not,
in fact, redundant: we injected a somewhat simplified definition and an
exact synonym that do not exist in RO. But since we did not inject such
simplified definitions and synonyms for other similar relations (e.g.
'existence starts during'), there is no reason to do it specifically for
'existence ends during'.

Related to #2453