Make 'supraorbital artery' supply blood to 'mucosa of frontal sinus'.
'supraorbital artery' (UBERON:0015160) supplies blood to the 'frontal
sinus' (UBERON:0001760), but 'frontal sinus' represents an immaterial
anatomical cavity, and is therefore incompatible with the range
restriction of the 'vessel supplies blood to' (RO:0020101) relation.

We replace the object of that relation to 'mucosa of frontal sinus' (a
material anatomical entity) instead.