Add 'intercostal element'.
The 'intercostal vein' (UBERON:0012197) and 'intercostal artery'
(UBERON:0005612) are said to drain and supply blood from and to the
'intercostal space' (UBERON:0012198).

In both cases, this violates the range restriction of the 'vessel drains
blood from' (RO:0020102) and 'vessel supplies blood to' (RO:0020101)
relations, because 'intercostal space' is an immaterial anatomical
entity.

We create a new term 'intercostal element' intended to refer to any
anatomical structure located within an intercostal space (e.g.
intercostal muscles, intercostal lymph nodes), and use the new term to
replace 'intercostal space' in the relationships with 'intercostal vein'
and 'intercostal arteries'.