Generate the bridge with SSLSO.
The SSLSO project provides its own mapping set, so we just need to fetch
it, then we can generate the bridge at the same time as all the other
bridges.

We do _not_ generate a distinct bridge for all the species present in
SSLSO, and we will not do that until/unless there is an explicit demand
for it. All bridging axioms to SSLSO terms are in a single bridge,
except for HsapDv and MmusDv terms (we need MmusDv as a separate bridge
to construct composite-mouse; there is no real reason I can think of to
have a separate HsapDv bridge, but we always had it, so I can already
hear people screaming if I dare remove it.)