Use batch version of the merge-species command.
Instead of calling the uberon:merge-species command repeatedly, once for
every species to merge, we call it only once, with a batch file listing
all the species for which a merge is required.

This removes some clutter from the Makefile, but most importantly this
also makes the whole operation much faster (from ~45min down to ~7min,
on my machine), because in batch mode the reasoner state is shared
between all merge operations -- we don't need to create a new reasoner
and have it reason over the ontology for every merge, which is what
takes the most time. The reasoner is initialised once at the beginning
of the first merge, and then it just needs to be kept updated for the
subsequent merge, which is much faster than creating a whole new
reasoner instance.