Use a "master" file for taxon-specific config data.
Currently, the information about which bridges to generate and how, and
which species to unfold in composite-metazoan and how, is dispersed in
two different places: in the bridges/bridges.rules.m4 source file to
generate the bridge, and the config/tax-merges.tsv to generate
composite-metazoan.

This commit proposes to make those config data more manageable by moving
them all to a single config/taxa.yaml file, from which we derive (using
a relatively simple Python script) both the SSSOM/T rule file and the
batch file that drives the compositing process.

Arguably, having the SSSOM/T ruleset being generated by a Python script
is more maintainer-friendly than having it generated by M4 macros, given
that there are likely many more ontology engineers that can read and
write Python than ontology engineers that can read and write M4 (which I
believe is a shame, as M4 is a powerful and lightweight tool that can do
great things when used well, but that's unfortunately beyond the point).