Re-define 'intercostal artery' and 'intercostal vein'.
We redefine 'intercostal artery' (respectively 'intercostal vein')
so that it is said to supply blood to (resp. drain blood from)
'intercostal muscle' rather than 'intercostal space'.

'intercostal space' is not suitable here as it is an immaterial entity,
which violates the range restriction of the 'vessel supplies blood to'
(resp. 'vessel drains blood from') relation.